patient-intake/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
annoy==1.17.3
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
coloredlogs==15.0.1
confection==0.1.5
cymem==2.0.8
daily-python==0.10.1
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.2.2
fastapi==0.115.4
fastembed==0.4.1
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
greenlet==3.1.1
grpcio==1.65.4
grpcio-tools==1.65.4
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-community==0.2.18
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langcodes==3.4.1
langsmith==0.1.141
language_data==1.2.0
lark==1.1.9
llama-cloud==0.1.4
llama-index==0.11.22
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.22
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.2.16
llama-index-multi-modal-llms-openai==0.2.3
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.13
loguru==0.7.2
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mmh3==4.1.0
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
nemoguardrails==0.10.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
nvidia-riva-client==2.16.0
onnx==1.17.0
onnxruntime==1.20.0
openai==1.54.3
orjson==3.10.11
packaging==24.1
pandas==2.2.3
pillow==10.3.0
pipecat-ai==0.0.39
preshed==3.0.9
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
py_rust_stemmers==0.1.3
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyloudnorm==0.1.1
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-deepcompare==1.0.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
scipy==1.14.1
shellingham==1.5.4
simpleeval==1.0.3
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
starlette==0.41.2
striprtf==0.0.26
sympy==1.13.3
tenacity==8.5.0
thinc==8.3.2
tiktoken==0.8.0
tokenizers==0.20.3
torch==2.3.1
torchaudio==2.3.1
tqdm==4.67.0
triton==2.3.1
typer==0.13.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
websockets==13.1
wrapt==1.16.0
yarl==1.17.1